Correct boundary condition ommitting routes that land on nginx parameter max
Summary:
Because you can have endless amount of app endpoints, we have to split the generated config into a couple of different location
directives, otherwise Nginx complains about a parameter being too long. This is done by setting a maximum amount of charaters Nginx
is happy with in a parameter and breaking off chunks from the larger allowed routes array and writing them to a file in sections of
the determined size. A boundary condition originated from this where the last route that landed on the determine max amount of
characters per parameter was not written to the config file. This has now been corrected.

Test Plan: Manually implemted the fix on dev console and made sure all endpoints are written to the generated config.

Reviewers: #devs, marco

Reviewed By: #devs, marco

Differential Revision: http://phabricator.thinkst.com:9000/D912