Correct parameter length check to include current location
Summary:
The check to ensure that the parameter length in the additional Nginx config does not exceed the
defined limit, did not include the current location to be added to the location directive parameter. This caused an edge
case that allowed the parameter length to sometimes exceed the limit. This has now been corrected.

Test Plan: Tested on a dev console.

Reviewers: #devs, azhar

Reviewed By: #devs, azhar

Subscribers: azhar

Differential Revision: http://phabricator.thinkst.com:9000/D951